refactored verified emails table
